Phase A3: move the shared decl vocabulary into core, breaking cbindgen->jnigen
The C adapter imported ConvertDecl/ConvertSpec/local_path_prefix from
the JNI adapter. Those are not JNI concepts -- they are the
language-neutral declaration vocabulary a build script writes -- so they
move down to api/core/decl.rs and both back-ends read them from there.
Two sibling adapters no longer reference each other.

Correcting the plan's boundary: the expand decls are shared, not
JNI-only. fun! expands to FunctionDecl and example-cbindgen's build.rs
already writes convert!(Millis).input(fun!(millis_from_raw)), so
FunctionDecl must reach the C side -- and it holds ExpandParamDecl /
ExpandReturnDecl, dragging ExpandDecl and FieldsDecl along. Harmless:
all four name only TypeKey, Origin, syn types and each other, never a
Kotlin or JNI type.

Moved macro bodies now say $crate::core::<T> rather than
$crate::lang::<T>, so they name where the type actually lives -- $crate
will later resolve to the crate that defines the macro. That forces the
7 moved names to be reachable at prebindgen::core, the one public-API
change here and purely additive; pub mod lang is byte-identical.

Zero golden drift, which is what proves this was a pure move. 535 lib +
34 doc + 20 + 10 passing; clippy and fmt clean on 1.85.0 and stable.
