+++ prebindgen/src/api/lang/jnigen/mod.rs
@@ -36,8 +36,6 @@
 pub mod jni;
 pub(crate) mod util;
 
-#[cfg(feature = "unstable-cbindgen")]
-pub(crate) use jni::ConvertSpec;
 pub use jni::{
     box_jboolean, box_jbyte, box_jchar, box_jdouble, box_jfloat, box_jint, box_jlong, box_jshort,
     decode_byte_array, decode_string, encode_byte_array, encode_string, matching, null_byte_array,
